Leonidas-Bot/python/requirements.txt
discord.py==2.1.0
python-dotenv
dpytest
requests==2.28.2
certifi==2022.12.7
charset-normalizer==3.0.1
grpcio==1.51.1
grpcio-tools==1.51.1
idna==3.4
protobuf==4.21.12
# PyAudio==0.2.13
# pydub==0.25.1
requests==2.28.2
urllib3==1.26.14